+++ views/budget/new.ejs
@@ -3,14 +3,14 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../css/style.css">
+    <link rel="stylesheet" href="/css/style.css">
     <title>CRUDdy Cash - New Budget</title>
 </head>
 <body>
     <%- include('../partials/nav.ejs') %>
     <main>
         <h1>New Budget</h1>
-        <form action="/budgets" method="post">
+        <form action="/user-budgets" method="post">
             <fieldset aria-labelledby="new-entry-legend">
                 <legend name="new-entry-legend">Create Budget</legend>
                 <div>
