+++ views/budget/show.ejs
@@ -38,7 +38,7 @@
                     <ul class="list">
                         <% incomeCategories.forEach( (category) => { %>
                             <li>
-                                <a href="/budgets/<%= budget._id %>/categories/<%= category._id %>"><%= category.name %></a>
+                                <a href="/user-budgets/<%= budget._id %>/categories/<%= category._id %>"><%= category.name %></a>
                             </li>
                         <% }) %>
                     </ul>
@@ -50,7 +50,7 @@
                     <ul class="list">
                         <% expenseCategories.forEach( (category) => { %>
                             <li>
-                                <a href="/budgets/<%= budget._id %>/categories/<%= category._id %>"><%= category.name %></a>
+                                <a href="/user-budgets/<%= budget._id %>/categories/<%= category._id %>"><%= category.name %></a>
                             </li>
                         <% }) %>
                     </ul>
@@ -58,16 +58,16 @@
             <% } %>
         <% } %>
         <div>
-            <form action="/budgets" method="get">
+            <form action="/user-budgets" method="get">
                 <button type="submit">Go Back</button>
             </form>
-            <form action="/budgets/<%= budget._id %>/categories/new" method="get">
+            <form action="/user-budgets/<%= budget._id %>/categories/new" method="get">
                 <button type="submit">Create Category</button>
             </form>
-            <form action="/budgets/<%= budget._id %>/edit" method="get">
+            <form action="/user-budgets/<%= budget._id %>/edit" method="get">
                 <button type="submit">Edit</button>
             </form>
-            <form action="/budgets/<%= budget._id %>?_method=delete" method="post">
+            <form action="/user-budgets/<%= budget._id %>?_method=delete" method="post">
                 <button type="submit">Delete</button>
             </form>
         </div>
