+++ views/partials/nav.ejs
@@ -1,5 +1,6 @@
 <nav>
     <a href="/dashboard">Home</a>
-    <a href="/budgets/new">New Budget</a>
+    <a href="/user-budgets">Budgets</a>
+    <a href="/user-budgets/new">New Budget</a>
     <a href="/auth/sign-out">Sign Out</a>
 </nav>
